frontend

+++ backend/Controllers/goalController.js
@@ -2,11 +2,13 @@ const asyncHandler = require('express-async-handler')
 
 const Goal = require('../Models/goalmodel')
 
+const User = require('../Models/userModel') // included to make sure that users can't update and delete each others goals
+
 //@desc  Get goals
 //@route GET //api/goals
 //@access private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id})
 
     res.status(200).json(goals)
 })
@@ -14,7 +16,7 @@ const getGoals = asyncHandler(async (req, res) => {
 //@desc  Set goals
 //@route POST //api/goals
 //@access private
-const seGoals = asyncHandler(async (req, res) => {
+const setGoals = asyncHandler(async (req, res) => {
     // error handler
     if(!req.body.text) {
         res.status(400)
@@ -22,7 +24,8 @@ const seGoals = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     })
     res.status(200).json(goal)
 })
@@ -35,6 +38,19 @@ const UpdateGoals = asyncHandler(async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    // making sure users do not update and delete each others goals
+    const user = await Goal.findById(req.user.id)
+    // Check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User mot foumd')
+    }
+    // make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.
@@ -54,6 +70,19 @@ const deleteGoals = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+    
+     // making sure users do not update and delete each others goals
+     const user = await Goal.findById(req.user.id)
+     // Check for user
+     if (!user) {
+         res.status(401)
+         throw new Error('User mot foumd')
+     }
+     // make sure the logged in user matches the goal user
+     if (goal.user.toString() !== user.id) {
+         res.status(401)
+         throw new Error('User not authorized')
+     }
 
     const deletedGoal = await Goal.deleteOne({ _id: req.params.id})
 
@@ -61,7 +90,7 @@ const deleteGoals = asyncHandler(async (req, res) => {
 })
 module.exports = {
     getGoals,
-    seGoals,
+    setGoals,
     UpdateGoals,
     deleteGoals,
 }
